requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
PySide6>=6.5.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0 